+++ packages/flame/lib/effects.dart
@@ -5,5 +5,5 @@ export 'src/effects/move_effect.dart';
 export 'src/effects/rotate_effect.dart';
 export 'src/effects/scale_effect.dart';
 export 'src/effects/sequence_effect.dart';
-
+export 'src/effects/size_effect.dart';
 export 'src/extensions/vector2.dart';