+++ packages/flame/test/game/camera_and_viewport_test.dart
@@ -124,7 +124,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0)', // camera translation
-          'translate(10.0, 10.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 10.0, 10.0, 0.0, 1.0)', // position component translate
           'translate(0.0, 0.0)', // position component anchor
         ],
       );
@@ -152,7 +152,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, -4.0, -4.0, 0.0, 1.0)', // camera translation
-          'translate(10.0, 10.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 10.0, 10.0, 0.0, 1.0)', // position component translate
           'translate(0.0, 0.0)', // position component anchor
         ],
       );
@@ -197,7 +197,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 40.0, 30.0, 0.0, 1.0)', // camera translation
-          'translate(10.0, 20.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 10.0, 20.0, 0.0, 1.0)', // position component translate
           'translate(-0.5, -0.5)', // position component anchor
         ],
         // result: 50 - w/2, 50 - h/2 (perfectly centered)
@@ -228,7 +228,7 @@ void main() {
         ),
         [
           'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, -550.0, -1920.0, 0.0, 1.0)', // camera translation
-          'translate(600.0, 2000.0)', // position component translation
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 600.0, 2000.0, 0.0, 1.0)', // position component translation
           'translate(-0.5, -0.5)', // position component anchor
         ],
         // result: 50 - w/2, 80 - h/2 (respects fractional relative offset)
@@ -320,7 +320,7 @@ void main() {
         ),
         [
           'transform(2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 1.0)', // camera translation and zoom
-          'translate(100.0, 100.0)', // position component
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 100.0, 100.0, 0.0, 1.0)', // position component translate
           'translate(-0.5, -0.5)', // anchor
         ],
       );
@@ -343,7 +343,7 @@ void main() {
         ),
         [
           'transform(2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 2.0, 0.0, 100.0, 100.0, 0.0, 1.0)', // camera translation and zoom
-          'translate(100.0, 100.0)', // position component
+          'transform(1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 100.0, 100.0, 0.0, 1.0)', // position component translate
           'translate(-0.5, -0.5)', // anchor
         ],
       );