+++ packages/flame/CHANGELOG.md
@@ -25,6 +25,9 @@
  - Fix truncated last frame in non-looping animations
  - Default size of `SpriteComponent` is `srcSize` instead of spritesheet size
  - Export test helper methods
+ - Rename `ScaleEffect` to `SizeEffect`
+ - Introduce `scale` on `PositionComponent`
+ - Add `ScaleEffect` that works on `scale` instead of `size`
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps